fix: Claims Retrieval Issue in SqlIdentityHubStore Due to Missing payloadFormat #107
